+++ inc/class.meta-tags.php
@@ -7,15 +7,10 @@ if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 /**
  * SEO Meta Tags
  */
-class MetaTags extends Singleton {
-  
-  private $prefix;
+class MetaTags {
 
-  public function __construct(SEO $seo) {
+  public function __construct() {
 
-    $this->prefix = $seo->prefix;
-
-    add_action('acf/init', [$this, 'acf_init'], 10);
     add_action('wp_head', [$this, 'wp_head'], 4);
     
     add_filter('pre_option_blogname', [$this, 'filter_blogname']);
@@ -90,7 +85,7 @@ class MetaTags extends Singleton {
    * @return void
    */
   private function get_term_title($term): string {
-    $value = $this->get_field('title', $term);
+    $value = seo()->get_field('title', $term);
     if( !$value ) $value = $term->name;
     return $value;
   }
@@ -105,11 +100,11 @@ class MetaTags extends Singleton {
     $value = false;
     $qo = get_queried_object();
     if( $qo instanceof \WP_Post ) {
-      $value = $this->get_field('description', $qo->ID);
+      $value = seo()->get_field('description', $qo->ID);
     } elseif( $qo instanceof \WP_Term ) {
-      $value = $this->get_field('description', $qo);
+      $value = seo()->get_field('description', $qo);
     }
-    if( !$value ) $value = $this->get_field('description', 'options');
+    if( !$value ) $value = seo()->get_field('description', 'options');
     return $value;
   }
 
@@ -120,142 +115,10 @@ class MetaTags extends Singleton {
    * @return string|bool
    */
   public function filter_blogname( $value ) {
-    if( $custom_title = $this->get_field('site_name', 'options') ) {
+    if( $custom_title = seo()->get_field('site_name', 'options') ) {
       $value = __($custom_title);
     }
     return $value;
-  }
-
-
-  /**
-   * Get options field
-   *
-   * @param string $name
-   * @return mixed
-   */
-  private function get_field($name, $post_id = 0, $format_value = true) {
-    return get_field("{$this->prefix}_{$name}", $post_id, $format_value);
-  }
-
-  /**
-   * Inits Admin
-   *
-   * @return void
-   */
-  public function acf_init() {
-    $this->add_admin_ui();
-  }
-
-  /**
-   * Add SEO Options Page
-   *
-   * @return void
-   */
-  private function add_admin_ui() {
-    
-    $fields = ['text', 'textarea', 'image'];
-    $field_types = [];
-    foreach( $fields as $key ) {
-      $field_type = $this->is_qtranslate_enabled() ? "qtranslate_$key" : $key;
-      $field_types[$key] = array_merge( acf_get_field_type($field_type)->defaults, ['type' => $field_type]);
-    }
-    
-    acf_add_options_page([
-      'page_title' => __('SEO Options', 'rhseo'),
-      'menu_slug' => "{$this->prefix}-options",
-      'position' => '59.6',
-      'icon_url' => 'dashicons-share'
-    ]);
-
-    $fields = [
-      array_merge($field_types['textarea'], [ 
-        'key' => "key_{$this->prefix}_description",
-        'name' => "{$this->prefix}_description",
-        'label' => $this->is_options_page() ? __('Site Description', 'rhseo') : __('Description', 'rhseo'),
-        'required' => false,
-        'instructions' => __('Optimal length: 50–160 characters', 'rhseo'),
-        'max' => 200,
-        'rows' => 2,
-        'acfml_multilingual' => true,
-      ]),
-      array_merge($field_types['image'], [ 
-        'key' => "key_{$this->prefix}_image",
-        'name' => "{$this->prefix}_image",
-        'label' => __('Preview Image', 'rhseo'),
-        'required' => false,
-        'instructions' => __('For services like Facebook or Twitter.', 'rhseo'),
-        'acfml_multilingual' => true,
-      ])
-    ];
-
-    if( $this->is_options_page() ) {
-      $fields = array_merge([
-        array_merge($field_types['text'], [ 
-          'key' => "key_{$this->prefix}_site_name",
-          'name' => "{$this->prefix}_site_name",
-          'label' => __('Site Name', 'rhseo'),
-          'acfml_multilingual' => true
-        ]),
-      ], $fields);
-    } else {
-      $fields = array_merge([
-        array_merge($field_types['text'], [ 
-          'key' => "key_{$this->prefix}_document_title",
-          'name' => "{$this->prefix}_document_title",
-          'instructions' => __('Optional. Overwrites the title.', 'rhseo'),
-          'label' => __('Document Title', 'rhseo'),
-          'acfml_multilingual' => true
-        ]),
-      ], $fields);
-    }
-
-    acf_add_local_field_group([
-      'key' => "group_{$this->prefix}_options",
-      'title' => $this->is_options_page() ? __('SEO Global Defaults', 'rhseo') : __('SEO', 'rhseo'),
-      'fields' => $fields,
-      'location' => [
-        [
-          [
-            'param' => 'options_page',
-            'operator' => '==',
-            'value' => "{$this->prefix}-options",
-          ],
-        ],
-        [
-          [
-            'param' => 'post_type',
-            'operator' => '==',
-            'value' => "post",
-          ],
-        ],
-        [
-          [
-            'param' => 'post_type',
-            'operator' => '!=',
-            'value' => "post",
-          ],
-        ],
-        [
-          [
-            'param' => 'taxonomy',
-            'operator' => '==',
-            'value' => "all",
-          ],
-        ],
-      ],
-      'menu_order' => 1000,
-      'position' => 'normal',
-      'active' => true,
-    ]);
-  }
-
-  /**
-   * Checks if qtranslate is enabled
-   *
-   * @return boolean
-   */
-  private function is_qtranslate_enabled() {
-    return defined('QTX_VERSION');
   }
 
   /**
@@ -278,15 +141,15 @@ class MetaTags extends Singleton {
 
     $qo = get_queried_object();
     if( $qo instanceof \WP_Post ) {
-      $value = $this->get_field("image", $qo->ID, false);
+      $value = seo()->get_field("image", $qo->ID, false);
       // ... then try post thumbnail
       if( !$value ) $value = get_post_thumbnail_id($qo->ID);
     } if( $qo instanceof \WP_Term ) {
-      $value = $this->get_field("image", $qo);
+      $value = seo()->get_field("image", $qo);
     }
 
     // finally try global setting
-    if( !$value ) $value = $this->get_field("image", 'options', false);
+    if( !$value ) $value = seo()->get_field("image", 'options', false);
 
     // bail early if empty
     if( empty($value) ) return $value;
@@ -296,15 +159,4 @@ class MetaTags extends Singleton {
     return $value;
   }
 
-  /**
-   * Detects the options page
-   *
-   * @return bool
-   */
-  public function is_options_page(): bool {
-    global $pagenow;
-    if( $pagenow !== 'admin.php' ) return false;
-    if( ($_GET['page'] ?? false) !== "{$this->prefix}-options") return false;
-    return true;
-  }
 }