+++ inc/class.yoast-compatibility.php
@@ -7,13 +7,10 @@ if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 /**
  * Compatibility for YOAST SEO
  */
-class Yoast_Compatibility extends Singleton {
-  
-  private $prefix;
+class Yoast_Compatibility {
 
-  public function __construct(SEO $seo) {
+  public function __construct() {
     add_action( 'plugins_loaded', [$this, 'disable_yoast_frontend'] );
-    add_action( 'plugins_loaded', [$this, 'disable_taxonomy_metaboxes'] );
     add_action( 'add_meta_boxes', [$this, 'meta_boxes'], 11 );
   }
 
@@ -23,15 +20,10 @@ class Yoast_Compatibility extends Singleton {
    * @return void
    */
   public function disable_yoast_frontend() {
-    if( is_admin() || !defined('WPSEO_VERSION') ) return;
+    if( is_admin() || !function_exists('\YoastSEO') ) return;
     $loader = \YoastSEO()->classes->get( \Yoast\WP\SEO\Loader::class );
     \remove_action( 'init', [ $loader, 'load_integrations' ] );
     \remove_action( 'rest_api_init', [ $loader, 'load_routes' ] );
-  }
-
-
-  public function disable_taxonomy_metaboxes() {
-    
   }
   
 
@@ -42,9 +34,7 @@ class Yoast_Compatibility extends Singleton {
    */
   public function meta_boxes() {
     remove_meta_box( 'wpseo_meta', null, 'normal' );
-    remove_meta_box( 'wpseo_meta', null, 'side' );
-
-    
+    remove_meta_box( 'wpseo_meta', null, 'side' );    
   }
 
 }