+++ changelog.md
@@ -1,5 +1,6 @@
 #### 1.0.2 (2021-02-02)
 
+- cleanup (#2f25438)
 - Redirect Attachments (#6216347)
 - Remove `users` from `wp-sitemap.xml` (#be5c576)
 - add German translation (#e793cd2)