+++ changelog.md
@@ -1,5 +1,6 @@
-#### 1.0.2 (2021-02-02)
+#### 1.0.2 (2021-02-03)
 
+- sitemap: `inject_meta_query_noindex` for posts and terms (#0f5d6b1)
 - Option to hide posts from search engines (robots, wp-sitemap.xml) (#e0ab2f2)
 - cleanup (#4925ee2)
 - Redirect Attachments (#6216347)